use PIDFD_GET_INFO ioctl when available
This way using pidfd_spawnp won't have to rely on procfs, avoiding an unpleasant edge-case
where the child is pawned but we can't get the pid.
And pidfd.{try_}wait will be able to return the exit status even after a process has been reaped.
At least on newer kernels.