Add AddrspacePtr for pointers to non-0 addrspaces
A new lang item `addrspace_ptr_type` is added. It must be a `struct AddrspacePtr<T: 'static, const ADDRSPACE: u32>`.
The content of this struct as written in the Rust source is ignored when computing the type layout. It is replaced by a pointer into address space `ADDRSPACE`.

As the content of the struct is replaced inside the compiler, it cannot be “looked at” in the Rust source.
To do something with the type (more than just holding it), rustc_intrinsics are used.

Intrinsics are added to to
- cast between `AddrspacePtr<T, addrspace::GENERIC>` and `*mut T`, this is just syntactical type conversion
- cast to a different type `T` (a no-op, just type conversion) or cast to a different address space (`addrspacecast` in llvm)
- convert to an integer (`ptrtoint` in llvm)
- add to the offset (`ptr::[wrapping_]offset`, `getelementptr` in llvm)
- and read/write from the pointer (`load/store` in llvm)

This should satisfy all basic needs for operations on pointers.

It is disallowed to deref an `AddrspacePtr` directly in Rust, read/write should be used instead.
Rustc still uses deref internally to implement read/write.