Correct hexagon "unwinder_private_data_size"
Discovered while porting libstd to hexagon-unknown-qurt: the
unwinder data size refers to the count of pointers in _Unwind_Exception but
when I put the value "35" intiially for hexagon linux, I incorrectly
considered the size of the exception context hexagon_thread_state_t
data structure.

Correct the value for hexagon linux and expand it to cover all
hexagon architecture instead.